requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Werkzeug==2.2.2
gunicorn==25.0.3
python-dotenv
git+https://github.com/MrPeterss/infosci-spark-client.git

duckdb>=0.10.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
requests>=2.31.0
tiktoken>=0.5.0
